refactor: split provider http config into pool and client configs
